+++ .babelrc
@@ -1,0 +1,5 @@
+{
+    "presets": [
+        "react-app"
+    ]
+}
