+++ src/App.js
@@ -1,23 +1,19 @@
-import logo from './logo.svg';
-import './App.css';
+
+import { TagInput } from "./lib";
+import { useState } from "react";
 
 function App() {
+  const [tags, setTags] = useState([]);
+
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div style={{ marginLeft: 200, marginTop: 40 }}>
+      <TagInput
+        inputStyle={{
+          width: "400px"
+        }}
+        placeholder={"Click Enter to add tags"}
+        tags={tags} setTags={setTags}
+        closeButtonStyle={{ color: 'blue' }} />
     </div>
   );
 }
