+++ src/lib/TagInput.css
@@ -1,0 +1,88 @@
+/* Style for the tag list */
+.tag-list {
+  display: flex;
+  flex-wrap: wrap;
+  margin-bottom: 0.5rem;
+}
+
+/* Style for each tag */
+.tag {
+  margin-top: 8px;
+  display: flex;
+  align-items: center;
+  margin-right: 0.5rem;
+  margin-bottom: 0.5rem;
+  padding: 0.375rem 0.75rem;
+  font-size: 0.875rem;
+  font-weight: 400;
+  line-height: 1.5;
+  color: #212529;
+  background-color: #e2e6ea;
+  border-radius: 0.25rem;
+}
+
+/* Style for the delete button */
+.delete {
+ margin-left: 4px;
+ cursor: pointer;
+ color: #212529;
+ font-weight: bold;
+}
+
+.delete:hover {
+  text-decoration: none;
+  color: #212529;
+  background-color: #c6d3e1;
+}
+
+.delete:focus-visible {
+  box-shadow: 0 0 0 0.25rem rgba(0, 123, 255, 0.25);
+}
+
+/* Style for the input field */
+.input-field {
+  display: block;
+  width: 100%;
+  padding: 0.375rem 0.75rem;
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.5;
+  color: #212529;
+  background-color: #fff;
+  border: 1px solid #ced4da;
+  border-radius: 0.25rem;
+  transition: border-color 0.15s ease-in-out, box-shadow 0.15s ease-in-out;
+}
+
+.input-field:focus-visible {
+  border-color: #007bff;
+  outline: none;
+  box-shadow: 0 0 0 0.25rem rgba(0, 123, 255, 0.25);
+}
+
+.input-field::placeholder {
+  color: #6c757d;
+  opacity: 1;
+}
+
+.input-field:disabled {
+  background-color: #e9ecef;
+  opacity: 1;
+}
+
+.input-field:read-only {
+  background-color: #e9ecef;
+}
+
+.input-field:invalid {
+  border-color: #dc3545;
+}
+
+.input-field:invalid:focus-visible {
+  box-shadow: 0 0 0 0.25rem rgba(220, 53, 69, 0.25);
+}
+
+.input-field:focus-visible + .tag-list {
+  border-color: #007bff;
+  box-shadow: 0 0 0 0.25rem rgba(0, 123, 255, 0.25);
+}
